chore(ast): move constants to seperate file
